requirements.txt
# Core dependencies
Flask==2.3.3
pyTelegramBotAPI==4.14.0
python-dotenv==1.1.0
gunicorn==23.0.0

# Additional debugging and development tools
colorlog==6.7.0
flask-cors==4.0.0

# Production server
waitress==3.0.1